criacao do endpoint vinculodepositopedecuba que cria um pe de cuba e vincula ao deposito fornecido

+++ src/main/java/com/scanai/api/domain/pedecuba/Pedecuba.java
@@ -24,7 +24,7 @@ public class Pedecuba {
     private Long fkpedecuba;
     private LocalDate datafimfermentacao;
     private LocalDate datainicio;
-    private int volume;
+    private float volume;
     private boolean valid;
 
     public Pedecuba(DadosCadastroPeDeCuba dados){
