+++ src/main/java/com/scanai/api/domain/pedecuba/dto/DadosDetalhamentoPeDeCuba.java
@@ -4,7 +4,7 @@ import com.scanai.api.domain.pedecuba.Pedecuba;
 
 import java.time.LocalDate;
 
-public record DadosDetalhamentoPeDeCuba(Long id, Long fkfuncionario, Long fkpedecuba, LocalDate datafimfermentacao, LocalDate datainicio, int volume, boolean valid) {
+public record DadosDetalhamentoPeDeCuba(Long id, Long fkfuncionario, Long fkpedecuba, LocalDate datafimfermentacao, LocalDate datainicio, float volume, boolean valid) {
     public DadosDetalhamentoPeDeCuba(Pedecuba pedecuba) {
         this(
                 pedecuba.getId(),
